refactor: Code segmentation
